pyproject.toml
[project]
name = "rest-api-modernized"
version = "0.1.0"
description = "Modernized REST API backend (FastAPI) for the security operations platform."
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "python-multipart",
  "sqlalchemy[asyncio]==2.0.37",
  "asyncpg",
  "alembic",
  "typer",

  # Auth (Keycloak-compatible OIDC / JWT validation)
  "httpx",
  "PyJWT",
  "cryptography",
]

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "ruff",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I", "UP"]
